requirements.txt
opencv-python-headless
streamlit
tensorflow-cpu